prima: Send DHCP stop Indication in disconnect case
Currently the driver honours the BTCOEXMODE received
from kernel, and does a DHCP start indication,
and does not allow scan to take place, because of the
presence of just one antenna in the PHY.
Becasue of this the scan never happens untill the
request for BTCOEXMODE is again received to disable the
COEX mode.
In the case where the BTCOEX mode to reset the flag is
not received at all (in case of platform/framework
entering bad state), the driver shall not allow scan
at all, which can lead to STA interface never coming
up.

Fix is to check whether the BTCOEX flag was set when
disconnect came, and if that's the case, reset the flag,
and send the DHCP stop indication to FW, so that it
can also resume the roam scan.

Change-Id: I23532eb2f1232abaa9f0014aebb119544b02f1ca
CRs-Fixed: 2429972