wlan: Delete hash entry only on add sta failure
In SAP mode, association is rejected since peer is in
deny list. In this scenario, del_sta is not sent to
firmware since STA context is not found in
limAssocRspTxCompleteCnf as hash entry is deleted in
limRejectAssociation. This will result in subsequent
add_sta for the same peer to fail.

Fix is to delete hash entry in limRejectAssociation
only on add_sta failure.

Change-Id: I7d8757afb47b26516a0ef6f598e5adbda5b1e86e
CRs-Fixed: 2433402