wlan: Send appropriate country code in regulatory hint
5GHz channels are not populated when changing band from
All to 2.4GHz and then changing band back to All from
2.4GHz. When kernel version is greater than 3.9 and when
country code is not '00', incorrect country code of 'OO'
is sent in regulatory_hint to update the channel information
in NV table. This will also cause the subsequent request,
to set the current country using regulatory hint, to fail.

Fix is to send appropriate country code of '00' instead
of 'OO' in regulatory hint.

Change-Id: I90681d383e38006730de68001a45a03e97d50cbb
CRs-Fixed: 2438099